+++ src/test/java/com/adventofcode/Day7Test.java
@@ -73,34 +73,34 @@ public class Day7Test {
      */
     @Test
     void testExample() {
-        assertThat(Intcode.thrusterSignal("3,15,3,16,1002,16,10,16,1,16,15,15,4,15,99,0,0", Arrays.asList(4, 3, 2, 1, 0))).isEqualTo(43210);
+        assertThat(Intcode.thrusterSignal("3,15,3,16,1002,16,10,16,1,16,15,15,4,15,99,0,0", Arrays.asList(4L, 3L, 2L, 1L, 0L))).isEqualTo(43210);
     }
 
     @Test
     void testMaxThrusterSignalExample1() {
-        Pair<List<Integer>, Integer> max = Intcode.maxThrusterSignal("3,15,3,16,1002,16,10,16,1,16,15,15,4,15,99,0,0", 0, 1, 2, 3, 4);
-        assertThat(max.getLeft()).containsExactly(4, 3, 2, 1, 0);
+        Pair<List<Long>, Long> max = Intcode.maxThrusterSignal("3,15,3,16,1002,16,10,16,1,16,15,15,4,15,99,0,0", 0L, 1L, 2L, 3L, 4L);
+        assertThat(max.getLeft()).containsExactly(4L, 3L, 2L, 1L, 0L);
         assertThat(max.getRight()).isEqualTo(43210);
     }
 
     @Test
     void testMaxThrusterSignalExample2() {
-        Pair<List<Integer>, Integer> max = Intcode.maxThrusterSignal("3,23,3,24,1002,24,10,24,1002,23,-1,23,101,5,23,23,1,24,23,23,4,23,99,0,0", 0, 1, 2, 3, 4);
-        assertThat(max.getLeft()).containsExactly(0, 1, 2, 3, 4);
-        assertThat(max.getRight()).isEqualTo(54321);
+        Pair<List<Long>, Long> max = Intcode.maxThrusterSignal("3,23,3,24,1002,24,10,24,1002,23,-1,23,101,5,23,23,1,24,23,23,4,23,99,0,0", 0L, 1L, 2L, 3L, 4L);
+        assertThat(max.getLeft()).containsExactly(0L, 1L, 2L, 3L, 4L);
+        assertThat(max.getRight()).isEqualTo(54321L);
     }
 
     @Test
     void testMaxThrusterSignalExample3() {
-        Pair<List<Integer>, Integer> max = Intcode.maxThrusterSignal("3,31,3,32,1002,32,10,32,1001,31,-2,31,1007,31,0,33,1002,33,7,33,1,33,31,31,1,32,31,31,4,31,99,0,0,0", 0, 1, 2, 3, 4);
-        assertThat(max.getLeft()).containsExactly(1, 0, 4, 3, 2);
+        Pair<List<Long>, Long> max = Intcode.maxThrusterSignal("3,31,3,32,1002,32,10,32,1001,31,-2,31,1007,31,0,33,1002,33,7,33,1,33,31,31,1,32,31,31,4,31,99,0,0,0", 0L, 1L, 2L, 3L, 4L);
+        assertThat(max.getLeft()).containsExactly(1L, 0L, 4L, 3L, 2L);
         assertThat(max.getRight()).isEqualTo(65210);
     }
 
     @Test
     void testInputPartOne() throws IOException {
-        Pair<List<Integer>, Integer> max = Intcode.maxThrusterSignal(FileUtils.readLine("/day/7/input"), 0, 1, 2, 3, 4);
-        assertThat(max.getLeft()).containsExactly(1, 3, 0, 4, 2);
+        Pair<List<Long>, Long> max = Intcode.maxThrusterSignal(FileUtils.readLine("/day/7/input"), 0L, 1L, 2L, 3L, 4L);
+        assertThat(max.getLeft()).containsExactly(1L, 3L, 0L, 4L, 2L);
         assertThat(max.getRight()).isEqualTo(24625);
     }
 
@@ -154,34 +154,34 @@ public class Day7Test {
      */
     @Test
     void testFeedbackLoop1() {
-        int thrusterSignal = Intcode.thrusterSignal("3,26,1001,26,-4,26,3,27,1002,27,2,27,1,27,26,27,4,27,1001,28,-1,28,1005,28,6,99,0,0,5", Arrays.asList(9, 8, 7, 6, 5));
+        long thrusterSignal = Intcode.thrusterSignal("3,26,1001,26,-4,26,3,27,1002,27,2,27,1,27,26,27,4,27,1001,28,-1,28,1005,28,6,99,0,0,5", Arrays.asList(9L, 8L, 7L, 6L, 5L));
         assertThat(thrusterSignal).isEqualTo(139629729);
     }
 
     @Test
     void testMaxFeedbackLoop1() {
-        Pair<List<Integer>, Integer> thrusterSignal = Intcode.maxThrusterSignal("3,26,1001,26,-4,26,3,27,1002,27,2,27,1,27,26,27,4,27,1001,28,-1,28,1005,28,6,99,0,0,5", 9, 8, 7, 6, 5);
+        Pair<List<Long>, Long> thrusterSignal = Intcode.maxThrusterSignal("3,26,1001,26,-4,26,3,27,1002,27,2,27,1,27,26,27,4,27,1001,28,-1,28,1005,28,6,99,0,0,5", 9L, 8L, 7L, 6L, 5L);
         assertThat(thrusterSignal.getRight()).isEqualTo(139629729);
-        assertThat(thrusterSignal.getLeft()).containsExactly(9, 8, 7, 6, 5);
+        assertThat(thrusterSignal.getLeft()).containsExactly(9L, 8L, 7L, 6L, 5L);
     }
 
     @Test
     void testFeedbackLoop2() {
-        int thrusterSignal = Intcode.thrusterSignal("3,52,1001,52,-5,52,3,53,1,52,56,54,1007,54,5,55,1005,55,26,1001,54,-5,54,1105,1,12,1,53,54,53,1008,54,0,55,1001,55,1,55,2,53,55,53,4,53,1001,56,-1,56,1005,56,6,99,0,0,0,0,10", Arrays.asList(9, 7, 8, 5, 6));
+        long thrusterSignal = Intcode.thrusterSignal("3,52,1001,52,-5,52,3,53,1,52,56,54,1007,54,5,55,1005,55,26,1001,54,-5,54,1105,1,12,1,53,54,53,1008,54,0,55,1001,55,1,55,2,53,55,53,4,53,1001,56,-1,56,1005,56,6,99,0,0,0,0,10", Arrays.asList(9L, 7L, 8L, 5L, 6L));
         assertThat(thrusterSignal).isEqualTo(18216);
     }
 
     @Test
     void testMaxFeedbackLoop2() {
-        Pair<List<Integer>, Integer> thrusterSignal = Intcode.maxThrusterSignal("3,52,1001,52,-5,52,3,53,1,52,56,54,1007,54,5,55,1005,55,26,1001,54,-5,54,1105,1,12,1,53,54,53,1008,54,0,55,1001,55,1,55,2,53,55,53,4,53,1001,56,-1,56,1005,56,6,99,0,0,0,0,10", 9, 8, 7, 6, 5);
+        Pair<List<Long>, Long> thrusterSignal = Intcode.maxThrusterSignal("3,52,1001,52,-5,52,3,53,1,52,56,54,1007,54,5,55,1005,55,26,1001,54,-5,54,1105,1,12,1,53,54,53,1008,54,0,55,1001,55,1,55,2,53,55,53,4,53,1001,56,-1,56,1005,56,6,99,0,0,0,0,10", 9L, 8L, 7L, 6L, 5L);
         assertThat(thrusterSignal.getRight()).isEqualTo(18216);
-        assertThat(thrusterSignal.getLeft()).containsExactly(9, 7, 8, 5, 6);
+        assertThat(thrusterSignal.getLeft()).containsExactly(9L, 7L, 8L, 5L, 6L);
     }
 
     @Test
     void testInputPartTwo() throws IOException {
-        Pair<List<Integer>, Integer> max = Intcode.maxThrusterSignal(FileUtils.readLine("/day/7/input"), 9, 8, 7, 6, 5);
-        assertThat(max.getLeft()).containsExactly(5, 8, 6, 9, 7);
+        Pair<List<Long>, Long> max = Intcode.maxThrusterSignal(FileUtils.readLine("/day/7/input"), 9L, 8L, 7L, 6L, 5L);
+        assertThat(max.getLeft()).containsExactly(5L, 8L, 6L, 9L, 7L);
         assertThat(max.getRight()).isEqualTo(36497698);
     }
 }
